back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Backend for MosProm HR platform"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "dishka>=1.7.2",
    "fastapi>=0.119.0",
    "pyjwt[crypto]>=2.10.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
